moving column update/delete dunit to base SplitClusterDUnitTestBase
now run the connector update/delete test in SplitClusterDUnitTest with Apache Spark
while the SplitSnappyClusterDUnitTest runs the embedded dunit

moved main test code to core ColumnUpdateDeleteTests for above

minor dependency updates to dtests
